pyproject.toml
[tool.poetry]
name = "clariad"
version = "0.1.0"
description = "AI Agent Ecosystem for Software Development"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.25"
langchain-core = "^0.3.60"
langgraph = "^0.4.5"
langfuse = "^2.60.5"
pgvector = "^0.4.1"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.4"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
aiohttp = "^3.11.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
ruff = "^0.11.10"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "RUF", "ISC"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
